Switch Ikala-Console to gradle
