--- Ikala-Console/.classpath
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<classpath>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8"/>
-	<classpathentry kind="src" path="src"/>
-	<classpathentry kind="src" path="test"/>
-	<classpathentry kind="lib" path="lib/testing/junit-4.12.jar">
-		<attributes>
-			<attribute name="javadoc_location" value="jar:platform:/resource/Ikala-Console/lib/info/junit-4.12-javadoc.jar!/"/>
-		</attributes>
-	</classpathentry>
-	<classpathentry kind="lib" path="lib/testing/hamcrest-all-1.3.jar"/>
-	<classpathentry kind="lib" path="lib/plugins/Ikala-Core-0.4.0.jar" sourcepath="lib/info/Ikala-Core-0.4.0-sources.jar">
-		<attributes>
-			<attribute name="javadoc_location" value="jar:platform:/resource/Ikala-Console/lib/info/Ikala-Core-0.4.0-javadoc.jar!/"/>
-		</attributes>
-	</classpathentry>
-	<classpathentry kind="output" path="bin"/>
-</classpath>
