--- Ikala-Console/src/com/ikalagaming/gui/console/events/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Events dispatched by the console.
- */
-package com.ikalagaming.gui.console.events;
