--- Ikala-Console/src/com/ikalagaming/gui/console/package-info.java
@@ -1,4 +1,0 @@
-/**
- * A console for interfacing with the engine, and associated classes.
- */
-package com.ikalagaming.gui.console;
