+++ Ikala-Console/src/main/java/com/ikalagaming/gui/console/events/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Events dispatched by the console.
+ */
+package com.ikalagaming.gui.console.events;
