+++ Ikala-Console/src/main/java/com/ikalagaming/gui/console/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * A console for interfacing with the engine, and associated classes.
+ */
+package com.ikalagaming.gui.console;
