+++ Ikala-Console/src/main/java/com/ikalagaming/gui/console/resources/Console.properties
@@ -1,0 +1,12 @@
+title=Console
+unknown_command=Unknown command
+try_cmd=For a list of available commands, type
+error_bad_location=Bad location
+not_enabled=Console is not enabled
+invalid_clipboard=Invalid clipboard contents
+missed_logs=Logs sent before console was enabled were skipped
+
+
+COMMAND_LIST_PLUGINS=plugins
+COMMAND_HELP=help
+COMMAND_VERSION=version
