+++ Ikala-Console/src/main/resources/plugin.yml
@@ -1,0 +1,5 @@
+name: Ikala-Console
+version: 0.1.2
+description: A console for interacting with the system
+authors: [maceris]
+main-class: com.ikalagaming.gui.console.ConsolePlugin
